refactor(config): centralize DNS routing validation
