+++ AGENTS.md
@@ -251,6 +251,7 @@ JA4 was verified on .70 with `/usr/local/bin/ja4probe` (`ja4probe`, source
     - `crates/honk-config/src/node.rs`, `src/config.rs`, `src/parser/tests.rs` — GroupPolicy serde, dae policy/default/errors, file loading, parser units: sections, groups, nested filters, DNS upstreams/routing, subscriptions, experimental.
     - `crates/honk-config/tests/example_configs.rs` — keeps `config.dae`, `config.min.dae`, `example.dae` parseable.
     - `crates/honk-config/tests/include.rs` — file-based dae include loading (glob order, nested paths, merge semantics, cycles, directory boundaries, and preservation of policy errors).
+    - `crates/honk-config/tests/dns_validation.rs` — named upstream references, original diagnostic paths, complete rule traversal, and shared effective-request precedence.
     - `crates/honk-config/tests/share_link.rs` — share-link parsing + config format round-trips.
     - `crates/honk-nfqueue/src/*` — NFQA parsing, netlink/nft encoding, exactly-once verdicts; the ignored `kernel_tests.rs` contract exercises the production queue/table in an isolated real netns.
     - `crates/honk-outbound/src/group/tests.rs`, `group/score/tests/`, and `group/udp_selection_repro_tests.rs` — selection semantics (Selector/URLTest/LoadBalance/Fallback/Score, nested groups, UDP exclusion), target attribution, decay and metric qualification, deterministic exploration, reporter lifecycle, reload sharing, and LRU safety.
@@ -294,6 +295,7 @@ Owned nftables names and firewall restrictions: `crates/honk-nfqueue`.
 - Group policy: omitting `policy` selects `Selector`; `policy: score` selects the always-compiled `Score` policy. Legacy `honk` is invalid. Score has no configuration knobs.
 - Built-ins: `honk-core` injects `direct`/`block`, not the parser (see `honk-config`'s `ensure_builtin_nodes()`). User nodes using those names or `NodeProtocol::Direct`/`Block` fail validation. `block` drops traffic.
 - `src/config.rs` — `Config` / `GlobalConfig` (~40 global fields), `from_file` / `to_file` / `validate`, JSON helpers. Idempotent `ensure_builtin_nodes()` injects `direct`/`block` with reserved `NodeProtocol::Direct`/`Block` and stable `DIRECT_NODE_ID`/`BLOCK_NODE_ID` UUIDs. **The crate never calls it**: `honk-core/src/lib.rs` calls at startup/SIGHUP; other consumers must call explicitly.
+  DNS upstream-reference checks live in `src/dns/validation.rs` and share `DnsRouting::request_source` with `effective_request`, preserving legacy diagnostic paths without converting rules during validation. Startup, SIGHUP, and public runtime reload call `Config::validate`; parse-only loaders do not.
 - Interface modes: absent `global.lan_interface` means no LAN hooks, never implicit `lo`; configured `wan_interface` hooks still proxy host TCP/UDP. Unresolved `auto` stays pending/fail-open until an IPv4 default route appears. On link/address/route events, `IfaceWatcher` attaches single-/dual-homed programs without restart, refreshes gateway-address must-direct rules through runtime configuration, immediately re-probes health-backed outbounds, and follows WAN bond slaves. WAN-only still requires `dae0`/`daens`, transparent listeners, cgroup hooks, bypass marks.
 - Health: `global { tcp_check_url, udp_check_dns, check_interval, check_tolerance }`. Selector leaves stay warm as specified under **Selector warm ownership**. `udp_warm_node_count: 0` disables the separate UDP warm set; positive values warm each group's top-N (≤3) reusable UDP leaves after every probe cycle. Dial modes: `ip` / `domain` / `domain+` / `domain++`.
 - Subscription recovery: `global.store_subscribe` defaults `true`, retaining the last valid raw response per request identity. Storage/fallback order and restore-before-refresh: `honk-core`'s `src/subscription.rs`. Process-scoped; reject SIGHUP changes as restart-required.
