+++ crates/honk-config/src/dns.rs
@@ -9,6 +9,8 @@ use url::Host;
 
 use crate::types::DnsProtocol;
 
+mod validation;
+
 /// Transports served by a standalone DNS bind endpoint.
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
 pub enum DnsBindTransport {
@@ -349,9 +351,9 @@ pub struct DnsUpstream {
 
 /// DNS routing configuration.
 ///
-/// Supports both the new request/response rules and the legacy flat
-/// `rules` + `fallback` format. When `request.rules` is empty (e.g.
-/// after serde from old JSON), `DnsRouter::new` converts legacy rules.
+/// Supports request/response rules and the legacy flat `rules` + `fallback`
+/// format. [`Self::effective_request`] applies their shared compatibility
+/// precedence for routing and policy identity.
 #[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
 pub struct DnsRouting {
     /// New-style request routing rules.
@@ -360,8 +362,8 @@ pub struct DnsRouting {
     /// New-style response routing rules.
     #[serde(default, skip_serializing_if = "DnsResponseRouting::is_default")]
     pub response: DnsResponseRouting,
-    /// LEGACY flat rules for old JSON/tests — converted in `DnsRouter::new`
-    /// when `request.rules` is empty.
+    /// Legacy flat rules; [`Self::effective_request`] selects them only when
+    /// new-style request rules are absent.
     #[serde(default)]
     pub rules: Vec<DnsRule>,
     /// Legacy fallback upstream name.
@@ -617,26 +619,32 @@ impl<'de> Deserialize<'de> for DnsResponseRouting {
     }
 }
 
+enum RequestSource {
+    Current,
+    Legacy,
+}
+
 impl DnsRouting {
     /// Select request routing with legacy compatibility precedence.
     pub fn effective_request(&self) -> Cow<'_, DnsRequestRouting> {
+        match self.request_source() {
+            RequestSource::Current => Cow::Borrowed(&self.request),
+            RequestSource::Legacy => Cow::Owned(self.convert_legacy_rules()),
+        }
+    }
+
+    fn request_source(&self) -> RequestSource {
         if !self.request.rules.is_empty() {
-            return Cow::Borrowed(&self.request);
+            return RequestSource::Current;
         }
-        if !self.rules.is_empty() {
-            return Cow::Owned(self.convert_legacy_rules());
+        if !self.rules.is_empty()
+            || (matches!(&self.request.fallback, DnsRequestAction::Upstream(name) if name == "default")
+                && !matches!(self.fallback.as_str(), "" | "upstream" | "default"))
+        {
+            RequestSource::Legacy
+        } else {
+            RequestSource::Current
         }
-        let uses_default = matches!(
-            &self.request.fallback,
-            DnsRequestAction::Upstream(name) if name == "default"
-        );
-        if uses_default && !matches!(self.fallback.as_str(), "" | "upstream" | "default") {
-            return Cow::Owned(DnsRequestRouting {
-                rules: vec![],
-                fallback: DnsRequestAction::Upstream(self.fallback.clone()),
-            });
-        }
-        Cow::Borrowed(&self.request)
     }
 
     /// Convert legacy rules into request rules.
