+++ doc/en/design/dns.md
@@ -25,6 +25,7 @@ Both ingress adapters use the same `DnsController`, current `DnsServiceProvider`
 
 - [`src/dns.rs`](../../../crates/honk-config/src/dns.rs) — `DnsConfig` (`bind`, `upstream`, `routing`, `strategy`, `cache`, `fixed_domain_ttl`). `DnsBindEndpoint` / `DnsBindTransport` / `DnsBindError` parse current-dae listeners with semantic equality; `DnsConfig::bind_endpoint` maps empty to disabled. Bind syntax: [Configuration](../configuration.md). `DnsUpstream`: name, address, `protocol: DnsProtocol`, `tls_server_name`, **`outbound: Option<String>`** dial-path proxy tag.
   Dae routing: first-match `DnsRequestRule`/`DnsResponseRule`, AND-ed negatable `DnsCond`s. Request: Qname/Qtype/Sip; response: Qname/Qtype/Upstream/Ip. `Sip` accepts mixed host/CIDR arguments, never response rules. Actions: Reject/AsIs/Accept/Upstream(name); legacy `rules`/`fallback` convert. `types.rs::DnsProtocol`: 6 variants—Udp, Tcp, Tls (DoT), Https (DoH), H3 (DoH3), Quic (DoQ). Only the dae parser populates request/response routing types; they deliberately sit outside serde.
+- [`src/dns/validation.rs`](../../../crates/honk-config/src/dns/validation.rs) owns named upstream-reference checks called by `Config::validate` at startup, SIGHUP, and public runtime reload. A private request-source selector in `DnsRouting` serves both validation and `effective_request`, preserving legacy field paths without allocating converted rules during validation. Parse-only loaders do not run whole-config validation.
 - DNS hosts use immutable, generation-pinned snapshots. Ordered `use_host` sources merge before request routing, cache lookup, or upstream exchange. Source syntax, precedence, and transactional SIGHUP behavior: [Configuration](../configuration.md).
 
 - [`src/dns/`](../../../crates/honk-core/src/dns/) — module ownership:
